refactor: organize dashboard route tests by widget
