+++ backend/tests/routes/dashboard/__init__.py
@@ -1,0 +1,1 @@
+"""Dashboard route tests"""
